requirements.txt
streamlit
transformers
torch
pytesseract
pdfplumber
joblib
openai
python-dotenv
Pillow
spacy
pandas